--- hw12_13_14_15_calendar/cmd/sender/internal/config/logger/logger.go
@@ -1,5 +1,0 @@
-package logger
-
-type Config struct {
-	Level string `env:"LEVEL" envDefault:"debug"`
-}
