+++ hw12_13_14_15_calendar/configs/.example.config.yaml
@@ -10,8 +10,6 @@ storage:
   user: ""
   password: ""
   db: "calendar"
-  timeouts:
-    read: 5s
 event:
   name: "Events"
 queue:
