+++ hw12_13_14_15_calendar/configs/settings/parser.go
@@ -1,4 +1,4 @@
-package parser
+package settings
 
 import (
 	"errors"
