+++ hw12_13_14_15_calendar/internal/app/sender/app.go
@@ -1,4 +1,4 @@
-package app
+package sender
 
 import (
 	"fmt"
